[shaders] Move shader wrapper code into a subfolder
